refactoring

+++ gift_certificates_system/persistence/src/main/java/com/epam/esm/dao/GiftCertificateDao.java
@@ -37,5 +37,5 @@ public interface GiftCertificateDao extends BaseDao<GiftCertificate> {
      * @param giftCertificateQueryParameters the gift certificate query parameters
      * @return the int
      */
-    int findTotalRecordsByQueryParameters(GiftCertificateQueryParameters giftCertificateQueryParameters);
+    long findTotalRecordsByQueryParameters(GiftCertificateQueryParameters giftCertificateQueryParameters);
 }
